Bruk Nais proxy i dev/prod. Sett proxy på RestTemplateBuilder  (Bean).
