+++ src/main/kotlin/no/nav/familie/ef/mottak/config/ApplicationConfig.kt
@@ -18,6 +18,7 @@ import org.springframework.boot.web.client.RestTemplateBuilder
 import org.springframework.boot.web.servlet.FilterRegistrationBean
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.ComponentScan
+import org.springframework.context.annotation.DependsOn
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories
 import org.springframework.web.client.RestOperations
 import springfox.documentation.swagger2.annotations.EnableSwagger2
@@ -44,15 +45,20 @@ class ApplicationConfig(@Value("\${application.name}")
     @Bean
     fun consumerIdClientInterceptor() = ConsumerIdClientInterceptor(consumerId = applicationName)
 
-
-    @Bean("restTemplate")
-    fun restTemplate(mdcInterceptor: MdcValuesPropagatingClientInterceptor,
-                     bearerTokenClientInterceptor: BearerTokenClientInterceptor): RestOperations {
+    @Bean("restTemplateBuilder")
+    fun restTemplateBuilder(): RestTemplateBuilder {
         return RestTemplateBuilder()
                 .additionalCustomizers(NaisProxyCustomizer())
-                .interceptors(mdcInterceptor, bearerTokenClientInterceptor)
-                .build()
+
     }
+
+    @Bean("restTemplate")
+    @DependsOn("restTemplateBuilder")
+    fun restTemplate(restTemplateBuilder: RestTemplateBuilder, mdcInterceptor: MdcValuesPropagatingClientInterceptor,
+                     bearerTokenClientInterceptor: BearerTokenClientInterceptor): RestOperations {
+        return restTemplateBuilder.interceptors(mdcInterceptor, bearerTokenClientInterceptor).build()
+    }
+
 
     @Bean
     fun kotlinModule(): KotlinModule = KotlinModule()
